[192-webreport] Implementing ReportObjectFactory - implemented comparison reports generator function

+++ jplag/src/main/java/de/jplag/reportingV2/reportobject/model/ComparisonReport.java
@@ -6,14 +6,14 @@ public class ComparisonReport {
 
 	private String first_submission_id;
 	private String second_submission_id;
-	private int match_percentage;
+	private float match_percentage;
 	private List<FilesOfSubmission> files_of_first_submission;
 	private List<FilesOfSubmission> files_of_second_submission;
 	private List<Match> matches;
 
 	public ComparisonReport(String first_submission_id,
 							String second_submission_id,
-							int match_percentage,
+							float match_percentage,
 							List<FilesOfSubmission> files_of_first_submission,
 							List<FilesOfSubmission> files_of_second_submission,
 							List<Match> matches) {
@@ -34,7 +34,7 @@ public class ComparisonReport {
 		return second_submission_id;
 	}
 
-	public int getMatch_percentage() {
+	public float getMatch_percentage() {
 		return match_percentage;
 	}
 
@@ -58,7 +58,7 @@ public class ComparisonReport {
 		this.second_submission_id = second_submission_id;
 	}
 
-	public void setMatch_percentage(int match_percentage) {
+	public void setMatch_percentage(float match_percentage) {
 		this.match_percentage = match_percentage;
 	}
